- Finished off TentativeApi implementation
