+++ Clash-Bot-Service-TS/controllers/TentativeController.js
@@ -1,0 +1,28 @@
+/**
+ * The TentativeController file is a very simple one, which does not need to be changed manually,
+ * unless there's a case where business logic routes the request to an entity which is not
+ * the service.
+ * The heavy lifting of the Controller item is done in Request.js - that is where request
+ * parameters are extracted and sent to the service, and where response is handled.
+ */
+
+const Controller = require('./Controller');
+const service = require('../services/TentativeService');
+
+const getTentativeDetails = async (request, response) => {
+  await Controller.handleRequest(request, response, service.getTentativeDetails);
+};
+
+const placePlayerOnTentative = async (request, response) => {
+  await Controller.handleRequest(request, response, service.placePlayerOnTentative);
+};
+
+const removePlayerFromTentative = async (request, response) => {
+  await Controller.handleRequest(request, response, service.removePlayerFromTentative);
+};
+
+module.exports = {
+  getTentativeDetails,
+  placePlayerOnTentative,
+  removePlayerFromTentative,
+};
