+++ Clash-Bot-Service-TS/controllers/index.js
@@ -1,11 +1,13 @@
 const SREController = require('./SREController');
 const TeamController = require('./TeamController');
+const TentativeController = require('./TentativeController');
 const TournamentController = require('./TournamentController');
 const UserController = require('./UserController');
 
 module.exports = {
   SREController,
   TeamController,
+  TentativeController,
   TournamentController,
   UserController,
 };
