+++ Clash-Bot-Service-TS/index.js
@@ -4,6 +4,7 @@ const ExpressServer = require('./expressServer');
 const clashUserDbImpl = require('./dao/clash-subscription-db-impl');
 const clashTimeDbImpl = require('./dao/clash-time-db-impl');
 const clashTeamsDb = require('./dao/clash-teams-db-impl');
+const clashTentativeDb = require('./dao/clash-tentative-db-impl');
 
 const launchServer = async () => {
   try {
@@ -11,6 +12,7 @@ const launchServer = async () => {
       clashUserDbImpl.initialize(),
       clashTimeDbImpl.initialize(),
       clashTeamsDb.initialize(),
+      clashTentativeDb.initialize(),
     ]);
 
     this.expressServer = new ExpressServer(config.URL_PORT, config.OPENAPI_YAML);
