+++ Clash-Bot-Service-TS/services/Service.js
@@ -1,12 +1,12 @@
-const logger = require("../logger");
+const logger = require('../logger');
 
 class Service {
   static rejectResponse(error, code = 500) {
-    return {error, code};
+    return { error, code };
   }
 
   static successResponse(payload, code = 200) {
-    return {payload, code};
+    return { payload, code };
   }
 
   static handleException({ error, reject, loggerContext = { class: 'Unknown', method: 'Unknown' } }) {
