+++ Clash-Bot-Service-TS/services/index.js
@@ -1,11 +1,13 @@
 const SREService = require('./SREService');
 const TeamService = require('./TeamService');
+const TentativeService = require('./TentativeService');
 const TournamentService = require('./TournamentService');
 const UserService = require('./UserService');
 
 module.exports = {
   SREService,
   TeamService,
+  TentativeService,
   TournamentService,
   UserService,
 };
