+++ ClashBot-WebApp-OpenAPI/open-api.yml
@@ -128,6 +128,7 @@ components:
       content:
         application/json:
           schema:
+            type: object
             required:
               - serverName
               - tournamentName
@@ -147,10 +148,10 @@ components:
                 $ref: '#/components/schemas/Player'
     PlayerCreateBody:
       description: All necessary parameters to create a new Player
-      required: true
       content:
         application/json:
           schema:
+            type: object
             required:
               - id
               - name
@@ -167,10 +168,10 @@ components:
                 type: string
     PlayerUpdateBody:
       description: Parameters to update a new Player
-      required: true
       content:
         application/json:
           schema:
+            type: object
             required:
               - id
             properties:
@@ -185,7 +186,6 @@ components:
                 type: string
     ChampionPostBody:
       description: A list of champions.
-      required: true
       content:
         application/json:
           schema:
@@ -237,6 +237,7 @@ components:
       content:
         application/json:
           schema:
+            type: object
             required:
               - serverName
               - playerId
@@ -377,7 +378,7 @@ paths:
       tags:
         - Team
       requestBody:
-        $ref: '#/components/schemas/TeamPatchPayload'
+        $ref: '#/components/requestBodies/TeamPatchBody'
       responses:
         '200':
           description: The Team was successfully updated.
@@ -459,7 +460,7 @@ paths:
           schema:
             type: string
       tags:
-        - Team
+        - Tentative
       responses:
         200:
           description: 'A tentative queue for a specific Tournament.'
@@ -477,7 +478,7 @@ paths:
       description: 'Places a player on the tentative queue for an upcoming Tournament.'
       operationId: placePlayerOnTentative
       tags:
-        - Team
+        - Tentative
       requestBody:
         $ref: '#/components/requestBodies/TentativeBody'
       responses:
@@ -495,7 +496,7 @@ paths:
       description: 'Remove a player from the tentative queue for an upcoming Tournament.'
       operationId: removePlayerFromTentative
       tags:
-        - Team
+        - Tentative
       requestBody:
         $ref: '#/components/requestBodies/TentativeBody'
       responses:
@@ -683,11 +684,7 @@ paths:
       tags:
         - User
       requestBody:
-        required: true
-        content:
-          application/json:
-            schema:
-              $ref: '#/components/schemas/ChampionUpdate'
+        $ref: '#/components/requestBodies/ChampionUpdateBody'
       responses:
         200:
           description: The updated list of preferred champions that the user has.
